Move datasource out of job configuration
