+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/MigrateReportingConfiguration.java
@@ -6,102 +6,51 @@ import org.opentestsystem.rdw.ingest.migrate.reporting.step.Warehouse;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecutionListener;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.BatchConfigurer;
-import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
-import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import javax.sql.DataSource;
 
 /**
  * MigrateReportingConfiguration.
- * Define the spring batch components to migrate data from the data warehouse to the reporting data mart.
+ * Spring Batch configuration for the migrate to reporting job.
  */
 @Configuration
 @EnableBatchProcessing
 public class MigrateReportingConfiguration {
-    //todo: move to constructor after resolving where to put batch tables, data source and transactions.
-    @Autowired
-    private JobBuilderFactory jobBuilderFactory;
 
-    @Autowired
-    private StepBuilderFactory stepBuilderFactory;
-
-
-    @Autowired
     private Reporting reportingStep;
-
-    @Autowired
     private Warehouse warehouseStep;
+    private JobBuilderFactory jobBuilderFactory;
+    private StepBuilderFactory stepBuilderFactory;
 
     public static final String stepReportingName = "stepReporting";
     public static final String stepWarehouseName = "stepWarehouse";
+
+    @Autowired
+    public MigrateReportingConfiguration(JobBuilderFactory jobBuilderFactory,
+                                         StepBuilderFactory stepBuilderFactory,
+                                         Reporting reportingStep,
+                                         Warehouse warehouseStep) {
+
+        this.jobBuilderFactory = jobBuilderFactory;
+        this.stepBuilderFactory = stepBuilderFactory;
+        this.reportingStep = reportingStep;
+        this.warehouseStep = warehouseStep;
+    }
 
     @Bean
     public JobExecutionListener listener() {
         return new MigrateJobCompletionListener();
     }
 
-    public PlatformTransactionManager transactionManager() {
-        return new ResourcelessTransactionManager();
-    }
-
-    @Primary
-    @Bean(name = "batchDataSource")
-    @ConfigurationProperties(prefix = "spring.batch_datasource")
-    public DataSource batchDataSource() {
-        return DataSourceBuilder
-                .create()
-                .build();
-    }
-
-    /**
-     * The jdbc template for reporting uses the same datasource as spring batch.
-     * This allows spring batch to control transactions to business data and spring batch tables.
-     * Prefix all tables with reporting or staging in sql when working on business data.
-     *
-     * @return NamedParameterJdbcTemplate
-     */
-    @Bean(name = "reportingJdbcTemplate")
-    public NamedParameterJdbcTemplate reportingJdbcTemplate() {
-        return new NamedParameterJdbcTemplate(batchDataSource());
-    }
-
-    @Bean
-    BatchConfigurer configurer(final DataSource dataSource) {
-        return new DefaultBatchConfigurer(dataSource);
-    }
-
-    @Bean
-    public JobRepository jobRepository() throws Exception {
-        JobRepositoryFactoryBean factory = new JobRepositoryFactoryBean();
-        factory.setDataSource(batchDataSource());
-        factory.setDatabaseType("MySQL");
-        factory.setTransactionManager(transactionManager());
-        factory.setIsolationLevelForCreate("ISOLATION_DEFAULT");
-        factory.setValidateTransactionState(false);
-        factory.afterPropertiesSet();
-        return factory.getObject();
-    }
-
     @Bean
     public Job migrateReportingJob() {
         try {
             return jobBuilderFactory.get("Migrate Reporting Job")
-                    .repository(jobRepository())
                     .incrementer(new RunIdIncrementer())
                     .listener(listener())
                     .flow(stepWarehouse())
