+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/configuration/DataSourceConfiguration.java
@@ -1,9 +1,12 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting.configuration;
 
+import org.springframework.batch.core.configuration.annotation.BatchConfigurer;
+import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.transaction.PlatformTransactionManager;
@@ -15,7 +18,47 @@ import javax.sql.DataSource;
  * Configuration for DataSources.
  */
 @Configuration
+@EnableTransactionManagement
 public class DataSourceConfiguration {
+
+    /**
+     * Data source for spring batch and reporting data mart with transaction manager.
+     *
+     * @return DataSource
+     */
+    @Primary
+    @Bean(name = "batchDataSource")
+    @ConfigurationProperties(prefix = "spring.batch_datasource")
+    public DataSource batchDataSource() {
+        return DataSourceBuilder
+                .create()
+                .build();
+    }
+
+    /**
+     * Spring Batch needs this if setting the data source.
+     * In addition to the primary being set.
+     * @return BatchConfigurer
+     */
+    @Bean
+    BatchConfigurer configurer() {
+        return new DefaultBatchConfigurer(batchDataSource());
+    }
+
+    @Bean
+    public PlatformTransactionManager reportingTransactionManager() {
+        return new DataSourceTransactionManager(batchDataSource());
+    }
+
+    /**
+     * The jdbc template used for reading and writing to the reporting data mart.
+     *
+     * @return DataSource
+     */
+    @Bean(name = "reportingJdbcTemplate")
+    public NamedParameterJdbcTemplate reportingJdbcTemplate() {
+        return new NamedParameterJdbcTemplate(batchDataSource());
+    }
 
     /**
      * Data source for the data warehouse with no transaction manager.
