+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/step/Warehouse.java
@@ -41,7 +41,7 @@ public class Warehouse extends StepTasklet {
                 .getJdbcOperations()
                 .queryForObject(findMaxExamImportId, Integer.class);
 
-        logger.info("=== query for reporting ===  Max exam import id: " + max);
+        logger.info("=== query for Warehouse ===  Max exam import id: " + max);
         return RepeatStatus.FINISHED;
     }
 }
