+++ migrate-reporting/src/main/resources/application.yml
@@ -17,3 +17,9 @@ spring:
     validationQuery: SELECT 1
     driverClassName: com.mysql.jdbc.Driver
     initialize: false
+
+logging:
+  level:
+    org.springframework.transaction: INFO
+    org.springframework.jdbc.core.JdbcTemplate: INFO
+    org.springframework.jdbc.datasource.DataSourceTransactionManager: INFO
